fix sorting example
